rename protection widgets
