+++ src/containers/datasets/index.tsx
@@ -80,6 +80,10 @@ import IUCNEcoregionInfo from 'containers/datasets/iucn-ecoregion/info.mdx';
 import IUCNEcoregionLayer from 'containers/datasets/iucn-ecoregion/layer';
 import IUCNEcoregionsMapLegend from 'containers/datasets/iucn-ecoregion/map-legend';
 import IUCNEcoregionWidget from 'containers/datasets/iucn-ecoregion/widget';
+import ProtectionInfo from 'containers/datasets/mangroves-in-protected-areas/info.mdx';
+import ProtectionLayer from 'containers/datasets/mangroves-in-protected-areas/layer';
+import ProtectionMapLegend from 'containers/datasets/mangroves-in-protected-areas/map-legend';
+import ProtectionWidget from 'containers/datasets/mangroves-in-protected-areas/widget';
 import NationalDashboardInfo from 'containers/datasets/national-dashboard/info.mdx';
 import NationalDashboardLayer from 'containers/datasets/national-dashboard/layer';
 import NationalDashboardMapLegend from 'containers/datasets/national-dashboard/map-legend';
@@ -88,10 +92,6 @@ import NetChangeInfo from 'containers/datasets/net-change/info.mdx';
 import NetChangeLayer from 'containers/datasets/net-change/layer';
 import NetChangeMapLegend from 'containers/datasets/net-change/map-legend';
 import NetChangeWidget from 'containers/datasets/net-change/widget';
-import ProtectionInfo from 'containers/datasets/protection/info.mdx';
-import ProtectionLayer from 'containers/datasets/protection/layer';
-import ProtectionMapLegend from 'containers/datasets/protection/map-legend';
-import ProtectionWidget from 'containers/datasets/protection/widget';
 import RestorationInfo from 'containers/datasets/restoration/info.mdx';
 import RestorationLayer from 'containers/datasets/restoration/layer';
 import RestorationMapLegend from 'containers/datasets/restoration/map-legend';
