+++ src/containers/datasets/mangroves-in-protected-areas/constants.tsx
@@ -1,0 +1,9 @@
+export const years = [1996, 2007, 2008, 2009, 2010, 2015, 2016, 2017, 2018, 2019, 2020];
+
+export const COLORS = {
+  '0% - 20%': 'rgba(207, 89, 126, 0.7)',
+  '20% - 40%': 'rgba(238, 180, 121, 0.7)',
+  '40% - 60%': 'rgba(233, 226, 156, 0.7)',
+  '60% - 80%': 'rgba(156, 203, 134, 0.7)',
+  '80% - 100%': 'rgba(0, 147, 146, 0.7)',
+};
