+++ src/containers/datasets/mangroves-in-protected-areas/info.mdx
