+++ src/containers/datasets/mangroves-in-protected-areas/types.d.ts
@@ -8,3 +8,7 @@ export type DataResponse = {
   data: Data;
   metadata: Metadata;
 };
+
+export type ProtectionType = {
+  location: string;
+};
