--- src/containers/datasets/protection/constants.tsx
@@ -1,1 +1,0 @@
-export const years = [1996, 2007, 2008, 2009, 2010, 2015, 2016, 2017, 2018, 2019, 2020];
